Refactor controller helpers
