+++ app/controllers/dashboard_helper.rb
@@ -1,27 +1,5 @@
-module ControllerHelper
+module DashboardHelper
     
-  def sort_by(attribute, list)
-    if attribute.eql? "location"
-      list.order(:city)
-    elsif attribute.eql? "interest"
-      list.order('level_of_interest DESC')
-    elsif attribute.eql? "company"
-      list.order(:company_id)
-    else
-      list
-    end
-  end
-
-  def list
-    if @location
-      Job.where(city: @location)
-    elsif @category
-      @category.jobs
-    else
-      set_company.jobs
-    end
-  end
-
   def format_interest
     low    = 0
     medium = 0
